version_2016_10_18(修改了WT模块中cloest二分法的实现)

+++ LTEV2X/WT_B.h
@@ -34,6 +34,6 @@ private:
 	Matrix readH(int VeUEIdx, int subCarrierIdx);//读取对应子载波的信道响应矩阵
 
 	int searchMCSLevelTable(double SINR);
-	int closest(std::vector<double> v, int target);
-	int closest2(std::vector<double> v, int target);
+	int closest(std::vector<double> v, double target);
+	int closest2(std::vector<double> v, double target);
 };
